Lighten the actions color
