+++ lib/pages/club-details/informations/club_contact.dart
@@ -4,7 +4,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:v34/commons/card/card.dart';
+import 'package:v34/commons/card/titled_card.dart';
 import 'package:v34/models/club.dart';
 import 'package:v34/utils/extensions.dart';
 import 'package:v34/utils/launch.dart';
@@ -38,45 +38,24 @@ class ClubContact extends StatelessWidget {
         actions: [
           if (club.phone.isNotNullAndNotEmpty())
             CardAction(
-              icon: FontAwesomeIcons.phone,
-              iconColor: Colors.green,
+              icon: Icons.phone,
               onTap: () {
                 launch("tel:${club.phone}");
               },
             ),
           if (club.phone.isNotNullAndNotEmpty())
             CardAction(
-              icon: FontAwesomeIcons.commentDots,
-              iconColor: Colors.orangeAccent,
+              icon: Icons.sms,
               onTap: () {
                 launch("sms:${club.phone}");
               },
             ),
           if (club.email.isNotNullAndNotEmpty())
             CardAction(
-              icon: FontAwesomeIcons.envelope,
-              iconColor: Colors.redAccent,
+              icon: Icons.mail,
               onTap: () {
                 launch("mailto:${club.email}");
               },
-            ),
-          if (club.websiteUrl.isNotNullAndNotEmpty())
-            CardAction(
-              icon: FontAwesomeIcons.globe,
-              iconColor: Colors.orangeAccent,
-              onTap: () => launchURL(club.websiteUrl),
-            ),
-          if (club.facebook.isNotNullAndNotEmpty())
-            CardAction(
-              icon: FontAwesomeIcons.facebook,
-              iconColor: Color(0xFF4064AD),
-              onTap: () => launchURL(club.facebook),
-            ),
-          if (club.twitter.isNotNullAndNotEmpty())
-            CardAction(
-              icon: FontAwesomeIcons.twitter,
-              iconColor: Color(0xFF1D9DEB),
-              onTap: () => launchURL(club.twitter),
             ),
         ]);
   }
